Retarget the guest build to rv64imac_zicsr/lp64
Companion to the core's RV64 conversion -- the emulator's ELF64 loader
can't read the old ELF32 binary at all, so the guest has to move too.
riscv.lds/start.S needed no changes (memory addresses are just numeric
literals; start.S's byte-at-a-time bss-zeroing loop is width-agnostic).

Also fixes two real pointer-truncation warnings in our own
libc_shim.c that LP64's 64-bit pointers turned from "coincidentally
fine because ilp32 pointers were already 32 bits" into actual bugs
waiting to happen: _sbrk's bump-allocator bounds check and _read's
WAD-offset-to-pointer cast both went through uint32_t/int instead of
uintptr_t. Our addresses are still small enough that the truncation
was harmless today, but there's no reason to leave a footgun sitting
there. (p_maputl.c's InterceptsOverrun has the same warning but is
vendored doomgeneric code, not ours to fix.)

Verified: rebuilt doomv-free.elf, confirmed via readelf it's genuinely
ELF64/RISC-V, and ran it live through the RV64 core -- correct
rendering/HUD/gameplay, trace log showing real RV64 instructions
(C.ADDIW, LD, SRLIW) executing mid-game, no crash.log.

+++ tools/doombuild/libc_shim.c
@@ -49,7 +49,7 @@ int _read(int fd, char *buf, int len)
 	if (iwad_fd_offset >= WAD_LENGTH) return 0;
 	if (iwad_fd_offset + (uint32_t)len > WAD_LENGTH) len = (int)(WAD_LENGTH - iwad_fd_offset);
 
-	const volatile uint8_t *src = (const volatile uint8_t *)(WAD_BASE + iwad_fd_offset);
+	const volatile uint8_t *src = (const volatile uint8_t *)(uintptr_t)(WAD_BASE + iwad_fd_offset);
 	for (int i = 0; i < len; i++) buf[i] = (char)src[i];
 	iwad_fd_offset += (uint32_t)len;
 
@@ -113,7 +113,7 @@ void *_sbrk(int incr)
 	if (heap_ptr == 0) heap_ptr = &end;
 
 	char *prev = heap_ptr;
-	if ((uint32_t)(heap_ptr + incr) > WAD_BASE) {
+	if ((uintptr_t)(heap_ptr + incr) > WAD_BASE) {
 		errno = ENOMEM;
 		return (void *)-1;
 	}
